t/tree-test-tables.dat

#data
<table><b>do</b> <b>do</b> do <b>do</b> <b>do</b></table>
#errors
no DOCTYPE
<b> in table
"do" in table
</b> in table
<b> in table
"do" in table
</b> in table
" do " in table
<b> in table
"do" in table
</b> in table
<b> in table
"do" in table
</b> in table
#note
See <http://krijnhoetmer.nl/irc-logs/whatwg/20090709#l-210>.
#document
| <html>
|   <head>
|   <body>
|     <b>
|       "do"
|     <b>
|       "do"
|     " do "
|     <b>
|       "do"
|     <b>
|       "do"
|     <table>
|       " "
|       " "
